# openapi.json
{"openapi":"3.1.0","info":{"title":"Second Brain API","description":"Backend for the Knowledge Synthesizer SPA","version":"1.0.0"},"paths":{"/api/session":{"get":{"tags":["core"],"summary":"Get Session Info","description":"Public session shape for the SPA — drives API base URL–agnostic clients.","operationId":"get_session_info_api_session_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/user/profile":{"get":{"tags":["core"],"summary":"Get User Profile","operationId":"get_user_profile_api_user_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/sources":{"get":{"tags":["core"],"summary":"List Sources","operationId":"list_sources_api_sources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["core"],"summary":"Add Source","operationId":"add_source_api_sources_post","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}},{"name":"source_type","in":"query","required":true,"schema":{"type":"string","title":"Source Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ingest/pdf":{"post":{"tags":["core"],"summary":"Ingest Pdf","operationId":"ingest_pdf_api_ingest_pdf_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ingest_pdf_api_ingest_pdf_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/search":{"post":{"tags":["core"],"summary":"Search","operationId":"search_api_search_post","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/digest":{"get":{"tags":["core"],"summary":"Get Morning Digest","operationId":"get_morning_digest_api_digest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/notifications":{"get":{"tags":["core"],"summary":"Get Notifications","operationId":"get_notifications_api_notifications_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/grants":{"get":{"tags":["ngo"],"summary":"List Grants","operationId":"list_grants_api_grants_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/grants/generate":{"post":{"tags":["ngo"],"summary":"Create Grant","operationId":"create_grant_api_grants_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrantRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/volunteers":{"get":{"tags":["ngo"],"summary":"List Schedules","operationId":"list_schedules_api_volunteers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["ngo"],"summary":"Create Volunteer Shift","operationId":"create_volunteer_shift_api_volunteers_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VolunteerShiftCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/volunteers/{schedule_id}":{"delete":{"tags":["ngo"],"summary":"Delete Volunteer Shift","operationId":"delete_volunteer_shift_api_volunteers__schedule_id__delete","parameters":[{"name":"schedule_id","in":"path","required":true,"schema":{"type":"integer","title":"Schedule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/volunteers/schedule":{"post":{"tags":["ngo"],"summary":"Auto Schedule","operationId":"auto_schedule_api_volunteers_schedule_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AutoScheduleRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/impact":{"get":{"tags":["ngo"],"summary":"List Impact Reports","operationId":"list_impact_reports_api_impact_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/impact/generate":{"post":{"tags":["ngo"],"summary":"Create Impact Report","operationId":"create_impact_report_api_impact_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImpactReportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/mentees":{"get":{"tags":["mentor"],"summary":"List Mentees","operationId":"list_mentees_api_mentees_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/mentees/{mentee_id}/brief":{"get":{"tags":["mentor"],"summary":"Get Mentee Brief","operationId":"get_mentee_brief_api_mentees__mentee_id__brief_get","parameters":[{"name":"mentee_id","in":"path","required":true,"schema":{"type":"integer","title":"Mentee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/mentees/{mentee_id}/session":{"post":{"tags":["mentor"],"summary":"Add Session Note","operationId":"add_session_note_api_mentees__mentee_id__session_post","parameters":[{"name":"mentee_id","in":"path","required":true,"schema":{"type":"integer","title":"Mentee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionNoteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects":{"get":{"tags":["agency"],"summary":"List Projects","operationId":"list_projects_api_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["agency"],"summary":"Add Project","operationId":"add_project_api_projects_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}/sync":{"get":{"tags":["agency"],"summary":"Sync Project Brief","operationId":"sync_project_brief_api_projects__project_id__sync_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards":{"get":{"tags":["study"],"summary":"Get Due Flashcards","operationId":"get_due_flashcards_api_flashcards_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/flashcards/all":{"get":{"tags":["study"],"summary":"Get All Flashcards","operationId":"get_all_flashcards_api_flashcards_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/flashcards/generate":{"post":{"tags":["study"],"summary":"Auto Generate Flashcards","operationId":"auto_generate_flashcards_api_flashcards_generate_post","parameters":[{"name":"document_id","in":"query","required":true,"schema":{"type":"integer","title":"Document Id"}},{"name":"deck_name","in":"query","required":true,"schema":{"type":"string","title":"Deck Name"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/{card_id}/review":{"put":{"tags":["study"],"summary":"Review Flashcard","operationId":"review_flashcard_api_flashcards__card_id__review_put","parameters":[{"name":"card_id","in":"path","required":true,"schema":{"type":"integer","title":"Card Id"}},{"name":"quality","in":"query","required":true,"schema":{"type":"integer","title":"Quality"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{provider}/connect":{"post":{"tags":["integrations"],"summary":"Connect Integration","description":"Initiates the OAuth connection flow.\nReturns the URL to redirect the user to.","operationId":"connect_integration_api__provider__connect_post","parameters":[{"name":"provider","in":"path","required":true,"schema":{"type":"string","title":"Provider"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/google/callback":{"get":{"tags":["integrations"],"summary":"Google Oauth Callback","description":"Handles the redirect back from Google.\nExchanges the authorization code for tokens and updates the database.\nRedirects to the frontend callback page.","operationId":"google_oauth_callback_api_google_callback_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"error","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/status":{"get":{"tags":["integrations"],"summary":"Get Integration Status","description":"Returns a dictionary of provider statuses for the current user.","operationId":"get_integration_status_api_status_get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{provider}/disconnect":{"post":{"tags":["integrations"],"summary":"Disconnect Integration","description":"Disconnects a provider and removes/updates the database record.","operationId":"disconnect_integration_api__provider__disconnect_post","parameters":[{"name":"provider","in":"path","required":true,"schema":{"type":"string","title":"Provider"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AutoScheduleRequest":{"properties":{"availability_data":{"type":"string","title":"Availability Data","default":""},"programs":{"type":"string","title":"Programs","default":""}},"type":"object","title":"AutoScheduleRequest"},"Body_ingest_pdf_api_ingest_pdf_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"source_id":{"type":"integer","title":"Source Id"}},"type":"object","required":["file","source_id"],"title":"Body_ingest_pdf_api_ingest_pdf_post"},"GrantRequest":{"properties":{"title":{"type":"string","title":"Title"},"deadline":{"type":"string","format":"date","title":"Deadline"},"org_info":{"type":"string","title":"Org Info"},"guidelines":{"type":"string","title":"Guidelines"}},"type":"object","required":["title","deadline","org_info","guidelines"],"title":"GrantRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImpactReportRequest":{"properties":{"period":{"type":"string","title":"Period"}},"type":"object","required":["period"],"title":"ImpactReportRequest"},"ProjectRequest":{"properties":{"name":{"type":"string","title":"Name"},"client_name":{"type":"string","title":"Client Name"},"tech_stack":{"type":"string","title":"Tech Stack"}},"type":"object","required":["name","client_name","tech_stack"],"title":"ProjectRequest"},"SessionNoteRequest":{"properties":{"summary":{"type":"string","title":"Summary"},"blockers":{"type":"string","title":"Blockers"}},"type":"object","required":["summary","blockers"],"title":"SessionNoteRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VolunteerShiftCreate":{"properties":{"volunteer_name":{"type":"string","title":"Volunteer Name"},"program_name":{"type":"string","title":"Program Name"},"shift_date":{"type":"string","format":"date","title":"Shift Date"},"shift_time":{"type":"string","title":"Shift Time","default":"9AM-1PM"},"status":{"type":"string","enum":["confirmed","pending","cancelled"],"title":"Status","default":"pending"}},"type":"object","required":["volunteer_name","program_name","shift_date"],"title":"VolunteerShiftCreate"}}}}